Joshuakinslow dev

+++ linux_clone.lua
@@ -167,7 +167,11 @@ if #args == 1 then
         file.write(request.readAll())
         file.close()
         request.close()
-        fs.move("/linux_clone.lua", "/bin/linux_clone.lua")
+        if (curDir == "") then
+            if fs.exists("/linux_clone.lua") then
+                fs.move("/linux_clone.lua", "/bin/linux_clone.lua")
+            end
+        end
         term.write("OS Reboot Required(y/n): ")
         local doReboot = read()
         if (doReboot == "y" or doReboot == "Y") then
@@ -272,7 +276,11 @@ if #args == 1 then
         file.write(request.readAll())
         file.close()
         request.close()
-        fs.move("/linux_clone.lua", "/bin/linux_clone.lua")
+        if (curDir == "") then
+            if fs.exists("/linux_clone.lua") then
+                fs.move("/linux_clone.lua", "/bin/linux_clone.lua")
+            end
+        end
         term.write("OS Reboot Required(y/n): ")
         local doReboot = read()
         if (doReboot == "y" or doReboot == "Y") then
@@ -317,7 +325,10 @@ if #args == 1 then
         end
         
         local data = request.readAll()
-        textutils.slowPrint(data)
+        print(data)
+        term.write("[")
+        textutils.slowWrite("################")
+        print("]")
         file.write(data)
         file.close()
         request.close()
